Update EditorA.xml and changed price to return a double instead of int

+++ app/src/main/java/com/viktorkhon/udacity_project_10_inventoryapp/InventoryAdapter.java
@@ -49,8 +49,8 @@ public class InventoryAdapter extends CursorAdapter {
 
         // Take out the value of each column and store it in a variable
         String name = cursor.getString(cursor.getColumnIndexOrThrow(InventoryEntry.COLUMN_NAME));
-        int quantity = cursor.getInt(cursor.getColumnIndexOrThrow(InventoryEntry.COLUMN_QTY));
-        int price = cursor.getInt(cursor.getColumnIndexOrThrow(InventoryEntry.COLUMN_PRICE));
+        double quantity = cursor.getDouble(cursor.getColumnIndexOrThrow(InventoryEntry.COLUMN_QTY));
+        double price = cursor.getDouble(cursor.getColumnIndexOrThrow(InventoryEntry.COLUMN_PRICE));
 
         // Assign a value extracted from the table to corresponding View in the ListView
         nameView.setText(name);
